♻️ refactor(types): restructure and remove redundant files
- reorganize import statements for clarity and consistency
- remove unused files: chat_completion_token_logprob, completion_choice, completion_usage
- integrate CompletionChoice and CompletionUsage into completions.py for simplified structure
