--- src/argoproxy/types/completion_choice.py
@@ -1,34 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from typing import Dict, List, Literal, Optional
-
-from pydantic import BaseModel
-
-__all__ = ["CompletionChoice", "Logprobs"]
-
-
-class Logprobs(BaseModel):
-    text_offset: Optional[List[int]] = None
-
-    token_logprobs: Optional[List[float]] = None
-
-    tokens: Optional[List[str]] = None
-
-    top_logprobs: Optional[List[Dict[str, float]]] = None
-
-
-class CompletionChoice(BaseModel):
-    finish_reason: Literal["stop", "length", "content_filter"]
-    """The reason the model stopped generating tokens.
-
-    This will be `stop` if the model hit a natural stop point or a provided stop
-    sequence, `length` if the maximum number of tokens specified in the request was
-    reached, or `content_filter` if content was omitted due to a flag from our
-    content filters.
-    """
-
-    index: int
-
-    logprobs: Optional[Logprobs] = None
-
-    text: str
